PLDM Runtime BMC R/R support: EECACHE write safety (commit 2/n)
BACKGROUND

Hostboot Runtime is capable of writing to Hostboot's EECACHE
virtual-PNOR partition. PNOR writes are performed via PLDM on BMC
systems. PNOR writes are performed one page (4096 bytes) at a time,
one PLDM write-request per page.

The BMC can undergo a reset/reload at any time (including during a
PNOR write) which may cause any PLDM request to time out or otherwise
fail. The runtime and IPL code therefore need to handle R/Rs while in
the middle of a PNOR write operation, particularly one that spans
multiple PLDM write requests.

THIS COMMIT

This commit adds failure handling for EECACHE writes on BMC
systems. The failure handling is mostly targeted at supporting errors
at runtime that are caused by BMC reset/reloads, but the IPL code
benefits some as well.

The bulk of this commit relates to a flag that has been added to
EECACHE record headers. This bit flag, called "pnor_write_in_progress,"
which indicates whether the last attempted PNOR write (which happens
via PLDM on BMC systems) succeeded completely. It works like this:

1. The flag is set to 1 when a PNOR write (flush) operation begins.
2. Pages from the modified EECACHE entries are written out to the BMC,
   one page at a time.
3. If all pages are written successfully, the flag is set back to 0.
   If any page write fails, the flag is not set back to 0.

At the beginning of each IPL, the flag is checked. If the flag is 1,
it means that an EECACHE write may have only partially succeeded, and
the data therein might be corrupt. This will cause the IPL code to
ignore the EECACHE entry and reread the cached EEPROM data from
hardware (and set the flag to 0).

IPL-time code will trigger an assertion and TI when a PNOR write fails
(runtime code does not). However, the flag system as described above
applies to IPL-time writes, so partial IPL-time EECACHE writes will be
caught by this commit as well.

COMPATIBILITY

This change is backward-compatible with existing EECACHE data. The
flag is set to 0 in existings installations which corresponds to the
value that means that a write is not in progress. A new EECACHE
version number isn't needed.

Change-Id: Icdc0a7b92c5ef9d230d2f1dad2c63467256c40ff
RTC: 299918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129177
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130221