PLDM Runtime BMC R/R support: GUARD error handling (commit 6/6)
This commit implements write-safety for the GUARD PNOR partition.

BACKGROUND

Hostboot Runtime is capable of writing to Hostboot's GUARD_DATA
virtual-PNOR partition. PNOR writes are performed via PLDM on BMC
systems. PNOR writes are performed one page (4096 bytes) at a time,
one PLDM write-request per page.

The BMC can undergo a reset/reload (R/R) at any time (including during
a PNOR write) which may cause any PLDM request to time out or
otherwise fail. The runtime and IPL code therefore need to handle R/Rs
while in the middle of a PNOR write operation, particularly one that
spans multiple PLDM write-requests.

THIS COMMIT

This commit adds failure handling for GUARD_DATA writes on BMC
systems. The failure handling is mostly targeted at supporting errors
at runtime that are caused by BMC reset/reloads, but the IPL code
benefits some as well.

The bulk of this commit relates to a flag that has been added to the
GUARD_DATA partition header. This bit flag, called
"GARD_FLAG_PNOR_WRITE_IN_PROGRESS," indicates whether the last
attempted PNOR write (which happens via PLDM on BMC systems) succeeded
completely. It works like this:

1. The flag starts with the value 1. (This is to retain backwards
   compatibility with existing GUARD partition data.)
2. The flag is set to 0 when a PNOR write (flush) operation begins.
3. Pages from the modified GUARD_DATA partition are written out to the
   BMC, one page at a time.
4. If all pages are written successfully, the flag is set back to 1.
   If any page write fails, the flag is not set back to 1.

At the beginning of each IPL, the flag will be checked by the BMC. If
the flag is 0, it means that a GUARD_DATA write may have only
partially succeeded, and the data therein might be corrupt. This will
cause the BMC to clear the GUARD partition and ignore its contents
(and set the flag back to 1).

IPL-time code will trigger an assertion and TI when a PNOR write fails
(runtime code does not). However, the flag system as described above
applies to IPL-time writes, so partial IPL-time GUARD_DATA writes will
be caught by this commit as well.

ADJUNCT REBOOTS

It's also possible for the HBRT adjunct to be rebooted. Normally, if a
PNOR write fails but then PNOR becomes accessible later, we can cope
by writing all of the unwritten pages to PNOR because we keep a
working copy in memory. It is ok if we lose power because the IPL will
restart and the BMC will clear the GUARD partition. However, it is
possible for the adjunct itself to be rebooted without IPLing from
scratch. If the adjunct is rebooted after a PNOR write fails, we lose
that in-memory copy and cannot recover a consistent view of the data
in PNOR. In this case, we leave the "write-in-progress" bit set to 0
and we do not attempt to write the GUARD partition at all (until the
next IPL), so that the BMC can clear the data on the next IPL.

MPIPL

The BMC should treat MPIPLs the same as normal IPLs for guard
purposes. Hostboot only applies core guards on MPIPLs, but it's not
terrible if some guards from the hypervisor (reported through HBRT)
are lost on an MPIPL.

COMPATIBILITY

This change is backward-compatible with existing GUARD partition
data. The flag is set to 1 in existing installations which corresponds
to the value that means that a write is not in progress. A new
GUARD_DATA version number isn't needed.

Change-Id: Ib48c3199ddaf427b6981c614b063ec74d2fa0d1f
RTC: 299918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129193
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>