Prevent Duplicate UD Callout Sections in Error Logs
This commit introduces changes to prevent duplicate User Details
callout sections from being created in error logs. This is achieved by
collecting the hashes of each UD binary data blob that is added to
each error log and comparing each new data blob to the existing list
of hashes. In case of a match, the UD section does not get added, since
it's a duplicate.

Change-Id: Ib53231b20ee6017e738316115c75b8d9ab04ef3d
RTC: 287397
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130101
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Ilya Smirnov <ismirno@us.ibm.com>