Add a callback to load and start the Power Management Complex (PMC)
Added a call to set up a Power Management Complex (PMC) callback to load and
start the PMC in the HBRT postInitCalls.  Use the host_callback interface
function to get a callback to fw_notify.  Add support in fw_notify to handle
the callback and call the load and start PMC function -
load_and_start_pm_complex.

The reason for this callback, is because the PHYP has a time limit for when
HBRT must respond, when doing an operation, in the init phase.
Unfortunately the PMC load and start operation can exceed this time limit
and has exceeded this time limit causing the PHYP to assume the HBRT is dead.
In lieu of updating the time limit, creating a callback to execute the PMC
load and start post init thereby avoiding the PHYP time limit altogether.

Change-Id: I307eaab4190576c6af8a1bccedf301f178c39591
CQ: SW546107
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129600
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>