Fill partial reads with known data during PLDM file io
The VMM/PNOR layer expects complete 4KB page reads but PLDM file
io does not guarantee that.  For all of the images produced by
a real build flow the data is padded to 4KB so there is no issue.
But if lids are patched in the lab that may not be the case.  This
is especially a problem for the RINGOVD as it is assumed to be at
least 2KB and when it isn't we end up with random CRCs during the
SBE update flow.

The change here will have the PNOR Resource Provider fill in the
remaining space with FF (same as an erased PNOR).  The provider
is responsible for responding to the page fault so it makes
logical sense to have the fill happen here, rather than at any
of the PLDM layers that have other scenarios.

Change-Id: Idac74dc4c2df1112a8e70bff5b9545b3c0796e03
CQ: SW547437
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130177
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>