Avoid infinite loops for missing TPMs
If the BMC doesn't report the TPM to us but we can see it, we
will deconfigure the target.  This results in a reconfig loop
because any istep with a deconfig will trigger a reconfig.

The change is to force an exception for host_discover_targets
for the same reason we exempt host_gard.  Both of these isteps
are establishing the baseline against which future deconfigs
are measured.  It isn't until after these steps complete that
we need to track deconfiguration.

Change-Id: I4f207faa92bcb195f0785bb9d4ad3d117c0c28e6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/129876
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>