PLDM Runtime BMC R/R support: Handle recursive error commit (commit 5/n)
On BMC systems, error logs are sent to the BMC via PLDM when they are
"committed." In certain situations, it is possible for the PLDM code
that is sending an error log to itself encounter an error condition
and commit the error, which will attempt to send the error via PLDM,
etc. If the PLDM error state is persistent, the code may create error
logs recursively ad infinitum.

Hostboot already has code to deal with errors in error log sends. When
a PLDM function returns an error to the error commit code, further
error transmissions are disabled. The problem is that new error logs
can be created and committed before any error is returned to the
caller the caller has a chance to disable the PLDM interface.

This commit adds code to preemptively temporarily disable the sending
of new error logs to the BMC while an error log is in the process of
being sent, breaking the recursive cycle, and re-enable sending them
after the commit is done. Now, if an error does happen during an error
commit, it will be able to reach the caller, and the caller can
permanently disable the BMC PLDM interface.

Change-Id: I9717e7802b89149be4308383461bfd226ad72eb3
RTC: 299918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130482
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>