Increase MCTP polling frequency in Hostboot Runtime
Hostboot Runtime uses a polling strategy to check for MCTP packets
that arrive in response to outgoing PLDM requests. Previously, the
algorithm would send a PLDM request, immediately check for a response,
and then begin polling with a frequency of one second. Since HBRT
performed the first check extremely quickly after the MCTP packet was
sent, it was very unlikely that the reply would have arrived by
then. This caused essentially every PLDM request to take at least one
second to finish, even in the cases when the requests only take e.g. 2
milliseconds to actually complete.

This commit reduces the MCTP polling delay to one millisecond so that
PLDM requests don't waste so much time.

Change-Id: I842c99f7fe05333629bf6ad716fde95160b979fa
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130210
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>