Implement Hostboot standalone workaround for istep 3.17 check
In istep 3.17, SBE checks to ensure that the fused mode bit in the
export control register matches the fused mode bit in the pervasive
control 0 register.  While the service processor configures the registers
correctly in combined simulations, the Hostboot standalone simulation needs to
execute a workaround to get both bits set correctly.

Change-Id: I79e190e6e89f35962036088d2ff330ef45531af7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130401
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>