Remove info log for inaccurate special wakeup calls
There is a path with ECO cores that always hits the coherency
check that we have for forcing wakeup off when the count is
not zero.  This isn't really a functional problem and making
a bunch of PELs adds a bunch of extra time to the HBRT boot.

Change-Id: Ic9ff62812def5cd4551506654fec299dbd876c63
CQ: SW548293
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130317
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>