Fix FAPINAME_POS for memport targets
The numbering change to handle more memport targets included a
bug in how we compute the value of ATTR_FAPINAME_POS.  This value
should follow the parent OCMB_CHIP, not the overall MEM_PORT
numbering.  This has the effect of breaking the attribute sync
over to Cronus that relies on these values to coordinate targets.

Change-Id: I49a909aa298eaa276cf4e6a716fa8c09b30fa4fc
CQ: SW548652
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130313
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>