Modify console output to display pres/func targets per proc
This commit changes the console display for the bitstrings that
represent present/fucntional targets from a system wide representation
to a per proc representation. This commit also makes a change in
the display for ECO cores. Before, ECO cores were counted with the
regular executable core bitstring. This was confusing as doing this
mixed the display with both ECO and non-ECO cores, making the system
seem like it had more executable core than in actuality and extra steps
needed to be taken, ie attribute dump, to determine which cores were
ECO or not. Now the display will have another entry for the functional
cache on the system which includes both ECO and non-ECO cores as well
as an entry for just non-ECO cores. This is so users can tell at a
glance which cores can be used for execution and if needed take the
difference between the cache cores and executable cores to determine
the ECO cores.

Example of new output:
12.63749|HWAS|---------------------------------
12.63750|HWAS|PRESENT>
12.65026|TARG|PROCS=C0000000
12.65033|TARG|PROC[00]:
12.65036|TARG|      CORE=F0000000 DIMM=FF00000000000000
12.65039|TARG|     CACHE=F0000000 OCMB=FF00
12.65047|TARG|PROC[01]:
12.65050|TARG|      CORE=F0000000 DIMM=0000000000000000
12.65053|TARG|     CACHE=F0000000 OCMB=0000
12.65058|HWAS|---------------------------------
12.69313|devtree|Syncing to BMC
15.60461|ISTEP  6. 7 - host_update_primary_tpm
15.61659|ISTEP  6. 8 - host_gard
16.88913|HWAS|---------------------------------
16.88914|HWAS|FUNCTIONAL>
16.90083|TARG|PROCS=C0000000
16.90090|TARG|PROC[00]:
16.90094|TARG|      CORE=F0000000 DIMM=FF00000000000000
16.90097|TARG|     CACHE=F0000000 OCMB=FF00
16.90104|TARG|PROC[01]:
16.90108|TARG|      CORE=F0000000 DIMM=0000000000000000
16.90111|TARG|     CACHE=F0000000 OCMB=0000
16.90119|HWAS|---------------------------------

Example of:
 - core 0 non-functional
 - core 1 as an ECO core

16.88913|HWAS|---------------------------------
16.88914|HWAS|FUNCTIONAL>
16.90083|TARG|PROCS=C0000000
16.90090|TARG|PROC[00]:
16.90094|TARG|      CORE=30000000 DIMM=FF00000000000000
16.90097|TARG|     CACHE=70000000 OCMB=FF00
16.90104|TARG|PROC[01]:
16.90108|TARG|      CORE=F0000000 DIMM=0000000000000000
16.90111|TARG|     CACHE=F0000000 OCMB=0000
16.90119|HWAS|---------------------------------

Change-Id: I75f5c19e4a567ef94f073b8a42061990b19dcee4
RTC:300983
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130406
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>