Add New OCC Callout Targets: Node, Proc
This commit adds new logic to callout node and proc when OCC indicates so.

If OCC detects a problem with a target, it sends the information to
HTMGT for it to make an error log and a callout. Currently, node,
proc, and VRM targets are not supported as OCC callouts. This commit
adds logic to return correct targets for node and proc (VRM to follow)
when they are requested by OCC.

Change-Id: I3c82bf996f2d698cdae8a593630d661c9bddc8b0
RTC: 304217
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130598
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>