New FCO algorithm which includes deconfig priorities for COREs
The old FCO algorithm was shoehorned in the restrictEC function which
caused the logic in that function to become fairly difficult to read and
maintain. The Chip Team requested that Hostboot follow a list of
priorities for FCO deconfig of cores in a proc but because of the
restrictEc function trying to do too many things that request could not
be fulfilled in a way that wouldn't make that function even more difficult
to understand.

This commit supplies a separate new FCO algorithm which includes priorty
core deconfig so that FCO and EC restriction can be uncoupled from one
another.

Change-Id: Ia7c24cf79dc8035c831b9cc1fe4c1845d631415e
RTC:300781
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131190
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>