Force Checkstop for some LPC Errors
For OpenBMC-based systems we want PRD to run on the BMC to handle
some LPC errors.  This hostboot commit will force a checkstop such
that PRD will get called to handle those errors rather than hostboot
performing a TI (Terminate Immediate).

The code to force this checkstop will go into both the Hostboot
Bootloader and the Hostboot Base Image.

Any LPC Errors that are not masked off will triggers this case.
The checkstop will be forced by setting the following FIR bit:
'EQ_L2_FIR[13] - NCU timed out waiting for powerbus to return data`.

NOTE that on FSP-based systems the errors seen in this defect already
result in a checkstop so it is ok for the logic to be common.  It is
unlikely we will hit this issue on FSP-based boxes.

Change-Id: Idc0cc4a71f4e9a1ae15329a35855d935e4d2f2a9
CQ:SW544579
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131168
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>