Deconfigure the IO SCM chip if paired PROC is deconfigured
This commit will deconfigure an IO SCM chip if the paired PROC is
deconfigured on a dual core module system.  If the IO SCM chip is
deconfigured no need to deconfigure the paired PROC to match.

In a Dual Chip Module with IO SCM chips, the even chip (PROC) is
paired with an odd chip (an IO SCM). The IO SCM has no cores/memory.

When the PROC is deconfigured then need to deconfigure the associated
IO SCM chip if not PHYP will have an issue and assert.

Created a class to do the handling of finding the pairs, grouping
them together and deconfiguring the IO SCM chip if it's paired
PROC is deconfigured.

Having the code encapsulated in a single class make it easy for
updates and to remove from the code base if the design changed.

RTC: 301401
CQ: SW546709
Change-Id: Ic5a9f0a6ae8483c64f6c0b78f3eee8e071f2c446
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131277
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>