Ignore VPD read errors in CRC verification code
Previously, VPD read errors were creating at least two error logs: one
for failing to cache the VPD, and one for failing to read the cache
when performing the VPD CRC check. This commit adds code to delete the
error log when checking the CRC if the error was caused by a VPD read
failure to avoid the redundant logs.

CQ: SW545401
Change-Id: I697473206265656e20616772656174207765656b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130957
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>