Reduce Severity of PDR Exchange Error
When HB completes the PDR exchange with BMC, we don't expect any
further PDR exchanges and delete the queue on which we listen for
incoming PDR exchange messages.

During one of the testcases, the tester removed and re-installed
a part (LCD display) during the boot (after we performed the
exchange) and HB produced PDR exchange errors, since we couldn't
process the message that came in on a queue that doesn't exist
anymore. This commit reduces the severities of these errors from
unrecoverable to informational in the case when the queue doesn't
exist.

Change-Id: I3977f4659ca1aa44863887b5148bf98264522490
CQ: SW549720
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131114
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>