Call p10_setup_ref_clock before p10_clock_test in istep 8
There is a scenario where setup_ref_clock might never be run
on any of the secondary processors. This commit adds a call to
the hwp in call_host_cbs_start() in istep 8.

ATTR_SYS_CLK_NE_TERMINATION_SITE may not be correct for
the secondary procs, get the correct value from root control
reg 7, bit 30, of the master proc and apply to secondary procs.

Change-Id: I3c8a28950b9b7f52c239bb4e445886c9b3e59644
CQ:SW544329
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131110
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>