Wrap mutex around entire indirect scom operation
There are instances where a single scom satellite can control
access to resources that are part of multiple logical units.
Therefore it is possible for a second write into the same scom
satellite before the previous one was complete could result in
one of the writes being lost or possibly corrupted.

Change-Id: I41dd6dd61af76c1437ebbc79c4ebcfc2d33a1fde
CQ: SW549402
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131057
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>