CXXTEST Expedite failure flows
During the parallel test case execution if any task crashes it may
leave unintended state (like interrupts turned off) so discontinue
running tests as soon as detected.  The ramifications of having the
interrupts turned off could lead to excessive times in simulation which
may disguise test environments for false failures, etc. (as an example).

If an expedited test case failure is detected and surfaced, there may however
be additional failures in the following test cases, but the payoff of catching
the first failure is deemed more efficient.   The success of the test
case run would ultimately fail anyway and need to be re-run, the time savings
justifies the design decision.

Change-Id: Ia80130e57dabc7331b96a109947df127eda9d121
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131115
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>