Return individual OCC status for get OCC State Sensor
The OCC sent in the get request was being ignored and so the function
was returning the same status for all OCCs (even if the processor was
garded, which means the OCC was not running).

This fix will return the individual status for the requested OCC.
It will also return DORMANT when the system is in safe mode.

Change-Id: I79f3d491a7a1d9091f9bc5f5acc5ce24846db6ad
CQ: SW549750
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131267
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>