Make the XML to HB converter use the specified name of enumeration types
Previously, for enumeration-type attributes, the XML to HB converter
script would always assume that the name of the underlying enumeration
type was the same as the name of the attribute. This meant that we
were unable to use the same enumeration type for multiple attributes,
for example.

This commit changes the script so that attributes are not limited to
using underlying enumeration types that have the same name as the
attribute itself.

Change-Id: I54ed9a5e46fd639770cefda24ea8bef235320f0a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131275
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Daniel M Crowell <dcrowell@us.ibm.com>