Clock failure support for eBMC systems with spare clocks
This commit implements support for using and deconfiguring spare
clocks on eBMC systems that have them, in concert with eBMC changes to
handle that side of the implementation.

The bulk of this commit concerns a new attribute,
ATTR_SYS_CLOCK_DECONFIG_STATE, that contains the functional state of
the system clocks/oscillators. This is the way that Hostboot
communicates to the BMC which clocks have failed. (Hostboot does not
model the clocks as targets, which is why we need a new attribute.)

At the beginning of the IPL, this attribute will be set to reflect the
CP_REFCLOCK_SELECT scratch register, which tells Hostboot which
oscillator to use. If a clock operation fails and an error log
deconfigures an oscillator, the attribute will be updated to reflect
which clocks have failed so far, and this attribute will be synced to
the BMC.

The BMC will then look at this attribute on reconfig loops to
determine which clocks it should attempt to use. All processors on the
system must use the same oscillator (because of hardware
requirements), so if no oscillators are available, the BMC will log an
error and fail the IPL.

Change-Id: I7468652070616c65206b696e67203a2d20646677
RTC: 295271
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131586
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>